common tags.ts for upsert function
