+++ app/api/opportunities/route.ts
@@ -12,6 +12,7 @@ import {
   sql,
 } from "drizzle-orm";
 import { opportunities, tags, user } from "@/lib/schema";
+import { upsertTagsAndGetIds } from "@/lib/tags";
 import { getCurrentUser } from "@/server/users";
 import { NextRequest, NextResponse } from "next/server";
 import { z } from "zod";
@@ -28,52 +29,6 @@ const opportunitySchema = z.object({
   endDate: z.string().optional(),
 });
 
-async function upsertTagsAndGetIds(tagNames: string[]): Promise<string[]> {
-  const normalized = Array.from(
-    new Set(tagNames.map((name) => name.trim()).filter(Boolean))
-  );
-
-  if (normalized.length === 0) return [];
-
-  const existing = await db
-    .select({ id: tags.id, name: tags.name })
-    .from(tags)
-    .where(inArray(tags.name, normalized));
-
-  const existingMap = new Map(existing.map((row) => [row.name, row.id]));
-  const missing = normalized.filter((name) => !existingMap.has(name));
-
-  let inserted: { id: string; name: string }[] = [];
-
-  if (missing.length > 0) {
-    inserted =
-      (await db
-        .insert(tags)
-        .values(missing.map((name) => ({ name })))
-        .onConflictDoNothing()
-        .returning({ id: tags.id, name: tags.name })) ?? [];
-
-    const stillMissing = missing.filter(
-      (name) => !inserted.some((row) => row.name === name)
-    );
-
-    if (stillMissing.length > 0) {
-      const fetched = await db
-        .select({ id: tags.id, name: tags.name })
-        .from(tags)
-        .where(inArray(tags.name, stillMissing));
-      inserted = inserted.concat(fetched);
-    }
-  }
-
-  const idLookup = new Map(
-    [...existing, ...inserted].map((row) => [row.name, row.id])
-  );
-
-  return normalized
-    .map((name) => idLookup.get(name))
-    .filter((id): id is string => Boolean(id));
-}
 
 export async function POST(req: NextRequest) {
   try {
